+++ core/upload.ts
@@ -1,4 +1,4 @@
-import type { Storage } from "../types.ts";
+import type { Entry, EntryMetadata, Storage } from "../types.ts";
 
 export interface UploadOptions {
   name: string;
@@ -18,5 +18,24 @@ export default class Upload {
     this.description = options.description;
     this.storage = options.storage;
     this.publicPath = options.publicPath;
+  }
+
+  async *[Symbol.asyncIterator](): AsyncGenerator<EntryMetadata> {
+    for await (const metadata of this.storage) {
+      yield metadata;
+    }
+  }
+
+  get(id: string): Entry {
+    return this.storage.get(id);
+  }
+
+  async delete(id: string): Promise<void> {
+    await this.storage.delete(id);
+  }
+
+  async rename(id: string, newId: string): Promise<void> {
+    const newName = this.storage.name(newId);
+    await this.storage.rename(id, newName);
   }
 }